HDPI-2983: move claimant information and reasons for possession tab details into separate shared builder
